pyproject.toml
[tool.poetry]
name = "langchain-test"
version = "0.1.0"
description = ""
authors = ["Kazuki Yokoi <kazuki11.yokoi11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.13"
openai = "^1.14.3"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
